requirements-dev.txt
black==20.8b1
eth-brownie>=1.11.0,<2.0.0
